Basic Stun State is done
Simple stun state can be tested through giving a temp script to the player and press "Q" the activate the raycast. Hiitting the enemy should play stun animation for 3 sec (can be modify in the inspector) but very buggy at the moment in term of animations. Plus weird animations kept playing as a consequence of stun state.