Partial Collectibles Implementation
Made some changes to the Player prefab in terms of UI. Changed some names of Actions in the InputManager and InputActions asset. Started working on the collectibles system. Added IsCollectible to InteractableBase and a check in InteractionManager to see if the object is a collectible. Added an event in InteractionManager that tells the UIManager that a collectible has been found. Started working on saving the data from a collectible and sending it to the Notebook UI.