Working tech demo for pedestals and a summoning circle
Object with a collider necessary for this.
Item in player hand, if it has the tag of the related puzzle, will be placed on top of the collider.
Once placed, if the item is the correct item expected by the pedestal, we can make the pedestal do something.
For collections of pedestal, the summoningcirclemanager checks if all pedestals have the correct item. When thy do, things happen.